magreport-backend/src/main/java/ru/magnit/magreportbackend/mapper/auth/RoleResponseMapper.java
package ru.magnit.magreportbackend.mapper.auth;

import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import ru.magnit.magreportbackend.domain.user.Role;
import ru.magnit.magreportbackend.dto.response.user.RoleResponse;
import ru.magnit.magreportbackend.mapper.Mapper;

@Service
@RequiredArgsConstructor
public class RoleResponseMapper implements Mapper<RoleResponse, Role> {

    @Override
    public RoleResponse from(Role source) {

        return new RoleResponse()
                .setId(source.getId())
                .setTypeId(source.getRoleType().getId())
                .setName(source.getName())
                .setDescription(source.getDescription())
                .setCreated(source.getCreatedDateTime())
                .setModified(source.getModifiedDateTime());
    }
}
